Patched bug in J-type detangling

+++ simulator/detangle.c
@@ -30,11 +30,11 @@ uint32_t detangle_rv32i(uint32_t word, ins_types_rv32i_t type) {
         case u_type:
             return word & 0xFFFFF000;
         case j_type:
-        {
-            uint8_t bit_20 = (word >> 31) & 0x1;
-            uint8_t bit_10_1 = (word >> 21) & 0x3FF;
-            uint8_t bit_11 = (word >> 20) & 0x1;
-            uint8_t bit_19_12 = (word >> 12) & 0xFF;
+        {   
+            uint32_t bit_20 = (word >> 31) & 0x1;
+            uint32_t bit_10_1 = (word >> 21) & 0x3FF;
+            uint32_t bit_11 = (word >> 20) & 0x1;
+            uint32_t bit_19_12 = (word >> 12) & 0xFF;
             return (bit_10_1 << 1) | (bit_11 << 11) | (bit_19_12 << 12) | (bit_20 << 20);
         }
         default:
